requirements.txt
python-telegram-bot[httpx]>=21.0,<22.0
redis>=5.0,<6.0
